requirements.txt
google-cloud-vision==2.3.1